fix(upgrade): widen st_mode portably for macOS (u16) vs Linux (u32)

+++ src/internal/upgrade/lock.rs
@@ -234,7 +234,9 @@ mod unix_impl {
             Ok(Some(match stat.st_mode & libc::S_IFMT {
                 libc::S_IFREG => EntryKind::Regular {
                     size: stat.st_size as u64,
-                    mode: stat.st_mode & 0o7777,
+                    // `st_mode` is u16 on macOS and u32 on Linux; widen
+                    // portably to the field type.
+                    mode: (stat.st_mode & 0o7777) as u32,
                 },
                 libc::S_IFDIR => EntryKind::Directory,
                 libc::S_IFLNK => EntryKind::Symlink,
